requirements.txt
astroid==2.2.5
autopep8==1.4.4
bleach==3.1.0
blinker==1.4
Bootstrap-Flask==1.0.10
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Faker==1.0.5
Flask==1.1.1
Flask-Caching==1.7.2
Flask-CKEditor==0.4.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
meta==1.0.2
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.1.0
redis==3.2.1
requests==2.22.0
sendgrid==6.0.5
six==1.12.0
SQLAlchemy==1.3.3
text-unidecode==1.2
typed-ast==1.3.1
Unidecode==1.0.23
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
